License Plates Tracker for Python
This python module provides functions and tools for working with license plates and its patterns,
like calculating the amount of combinations possible for a specific type of plate,
or getting the nth license plate for a given pattern.

Standard plate patterns for Europe countries, US states and Latin America countries are
provided within the module.